+++ src/redux/configureStore.js
@@ -1,16 +1,18 @@
-import { createBrowserHistory } from 'history';
+import { createHashHistory } from 'history';
 import { applyMiddleware, compose, createStore } from 'redux';
 import { routerMiddleware } from 'connected-react-router';
 import thunk from 'redux-thunk';
 import createRootReducer from './reducers';
+import customMiddlewares from './configureMiddlewares';
 
-export const history = createBrowserHistory();
+export const history = createHashHistory();
 
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 const enhancer = composeEnhancers(
   applyMiddleware(
     routerMiddleware(history),
-    thunk
+    thunk,
+    ...customMiddlewares
   )
 );