+++ src/redux/reducers/stock/actions.js
@@ -1,4 +1,21 @@
 import axios from 'axios';
+import { BASE_CURRENCY } from '../../../utils/constants';
+import { snackActions } from '../../../utils/SnackbarUtils';
+
+const BASE_URL = 'https://api.coingecko.com/api/v3/coins/markets';
+const NO_CONNECTION_SNACKBAR_KEY = 'noConnection';
+const AVAILABLE_STOCKS = [
+  'bitcoin',
+  'ethereum',
+  'cardano',
+  'dogecoin',
+  'ripple',
+  'polkadot',
+  'uniswap',
+  'chainlink',
+  'litecoin',
+  'stellar'
+];
 
 export const STOCK_GET_INIT = 'stock/get/init';
 export const STOCK_GET_SUCCESS = 'stock/get/success';
@@ -16,23 +33,9 @@ export const getStockSuccess = (stockData) => ({
   payload: { stockData }
 });
 
-const AVAILABLE_STOCKS = [
-  'bitcoin',
-  'ethereum',
-  'cardano',
-  'dogecoin',
-  'ripple',
-  'polkadot',
-  'uniswap',
-  'chainlink',
-  'litecoin',
-  'stellar'
-];
-const BASE_STOCK = 'usd';
-const BASE_URL = 'https://api.coingecko.com/api/v3/coins/markets';
-
 const fetchStockData = async () => {
-  const url = `${BASE_URL}?vs_currency=${BASE_STOCK}&ids=${AVAILABLE_STOCKS.join(',')}`;
+  const cacheBuster = Math.round(new Date().getTime() / 1000);
+  const url = `${BASE_URL}?vs_currency=${BASE_CURRENCY}&ids=${AVAILABLE_STOCKS.join(',')}&cb=${cacheBuster}`;
   const { data } = await axios.get(url);
   return data;
 };
@@ -42,7 +45,14 @@ export const getTickers = () => async (dispatch) => {
   try {
     const data = await fetchStockData();
     dispatch(getStockSuccess(data));
+    snackActions.dismiss(NO_CONNECTION_SNACKBAR_KEY);
   } catch (error) {
-    dispatch(getStockFail(error));
+    dispatch(getStockFail(error.message));
+    snackActions.toast(error.message, {
+      variant: 'error',
+      preventDuplicate: true,
+      persist: true,
+      key: NO_CONNECTION_SNACKBAR_KEY
+    });
   }
 };